dagmm_training/requirements.txt
# DAGMM & DeepLog Training Requirements
# H100 GPU 최적화

torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.60.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
